revision_binder добавил возможность создавать экраны без reducer и state. Разделил по пакетам
