+++ mvi-core/src/main/kotlin/ru/surfstudio/mvi/vm/compose/FoundationMviCompose.kt
@@ -1,0 +1,9 @@
+package ru.surfstudio.mvi.vm.compose
+
+import ru.surfstudio.mvi.core.event.Event
+
+
+/** Helpful interface for emitting events from @Composable */
+fun interface ComposedViewContext<E : Event> {
+    fun emit(event: E)
+}
