+++ mvi-core/src/main/kotlin/ru/surfstudio/mvi/vm/compose/MviRenders.kt
@@ -1,0 +1,43 @@
+package ru.surfstudio.mvi.vm.compose
+
+import android.annotation.SuppressLint
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.rememberCoroutineScope
+import kotlinx.coroutines.launch
+import ru.surfstudio.mvi.core.event.Event
+import ru.surfstudio.mvi.vm.MviViewModel
+import ru.surfstudio.mvi.vm.MviViewModelWithoutState
+
+/** Syntax sugar fun for convenient binding in @Composable with MVI */
+@SuppressLint("ComposableNaming")
+@Composable
+infix fun <E : Event> MviViewModelWithoutState<E>.renders(
+    render: @Composable ComposedViewContext<E>.() -> Unit
+) {
+    val scope = rememberCoroutineScope()
+
+    ComposedViewContext<E> { event ->
+        scope.launch {
+            hub.emit(event)
+        }
+    }.render()
+}
+
+/** Syntax sugar fun for convenient binding in @Composable with MVI */
+@SuppressLint("ComposableNaming")
+@Composable
+infix fun <S : Any, E : Event> MviViewModel<S, E>.renders(
+    render: @Composable ComposedViewContext<E>.(S) -> Unit
+) {
+    val state by state.observeState().collectAsState(initial = state.currentState)
+    val scope = rememberCoroutineScope()
+
+    val composedViewContext = ComposedViewContext<E> { event ->
+        scope.launch {
+            hub.emit(event)
+        }
+    }
+    composedViewContext.render(state)
+}
